refactor: remove useSetupPasskey and integrate setupPasskey into usePasskey store
