+++ src/stores/passkey/usePasskey.ts
@@ -72,6 +72,30 @@ export const usePasskeyStore = defineStore(
 			}
 		}
 
+		async function setupPasskey() {
+			// Only auto recognize as logged in if the selected credential is stored in local storage and is full
+			// Because the full credential has username,
+			// the user can know which credential should be selected when signing transaction
+			if (selectedCredentialId.value && !isFullCredential.value) {
+				resetCredentialId()
+			}
+
+			// Check if passkey is supported by the browser
+			try {
+				if (
+					window.PublicKeyCredential &&
+					typeof PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable === 'function'
+				) {
+					isPasskeySupported.value = await PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable()
+				} else {
+					isPasskeySupported.value = false
+				}
+			} catch (err) {
+				isPasskeySupported.value = false
+				throw new Error(`Error checking passkey support: ${err}`)
+			}
+		}
+
 		return {
 			isPasskeySupported,
 			storedCredentials,
@@ -85,6 +109,7 @@ export const usePasskeyStore = defineStore(
 			resetCredentialId,
 			isFullCredential,
 			selectedCredentialDisplay,
+			setupPasskey,
 		}
 	},
 	{
